iox-#1196 Fix useless cast warnings

+++ iceoryx_hoofs/source/posix_wrapper/message_queue.cpp
@@ -177,8 +177,8 @@ cxx::expected<IpcChannelError> MessageQueue::destroy() noexcept
 
 cxx::expected<IpcChannelError> MessageQueue::send(const std::string& msg) const noexcept
 {
-    const size_t messageSize = static_cast<size_t>(msg.size()) + NULL_TERMINATOR_SIZE;
-    if (messageSize > static_cast<size_t>(m_attributes.mq_msgsize))
+    const uint64_t messageSize = msg.size() + NULL_TERMINATOR_SIZE;
+    if (messageSize > static_cast<uint64_t>(m_attributes.mq_msgsize))
     {
         return cxx::error<IpcChannelError>(IpcChannelError::MESSAGE_TOO_LONG);
     }
@@ -303,8 +303,8 @@ cxx::expected<std::string, IpcChannelError> MessageQueue::timedReceive(const uni
 cxx::expected<IpcChannelError> MessageQueue::timedSend(const std::string& msg,
                                                        const units::Duration& timeout) const noexcept
 {
-    const size_t messageSize = static_cast<size_t>(msg.size()) + NULL_TERMINATOR_SIZE;
-    if (messageSize > static_cast<size_t>(m_attributes.mq_msgsize))
+    const uint64_t messageSize = msg.size() + NULL_TERMINATOR_SIZE;
+    if (messageSize > static_cast<uint64_t>(m_attributes.mq_msgsize))
     {
         std::cerr << "the message \"" << msg << "\" which should be sent to the message queue \"" << m_name
                   << "\" is too long" << std::endl;
